requirements.txt
asn1crypto==0.24.0
bayesianpy==0.1
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cryptography==2.6.1
dask==0.18.0
dill==0.2.8.2
distributed==1.25.2
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.14
HeapDict==1.0.0
idna==2.8
jmespath==0.9.4
JPype1==0.6.3
locket==0.2.0
mkl-fft==1.0.10
mkl-random==1.0.2
msgpack==0.6.0
multiprocess==0.70.6.1
numpy==1.15.4
pandas==0.23.4
partd==0.3.9
pathos==0.2.2.1
pox==0.2.4
ppft==1.6.4.8
psutil==5.4.8
pycparser==2.19
pyOpenSSL==19.0.0
pypiwin32==223
pyreadline==2.1
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
pywin32==224
PyYAML==3.13
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==1.2.16
tblib==1.3.2
toolz==0.9.0
tornado==5.1.1
urllib3==1.24.1
websocket-client==0.56.0
win-inet-pton==1.1.0
wincertstore==0.2
zict==0.1.3
